[story/VSPC-213] refactor
