+++ vspace/src/main/java/edu/asu/diging/vspace/core/services/IExhibitionManager.java
@@ -34,4 +34,6 @@ public interface IExhibitionManager {
 
     IExhibition getStartExhibition();
 
+    void updateExhibitionLanguages(Exhibition exhibition, List<String> languages, String defaultLanguage);
+
 }
